# swagger-api.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Standup Comedy Connect API",
    "description": "API pour la plateforme de mise en relation entre humoristes et organisateurs d'événements de stand-up comedy",
    "version": "1.0.0",
    "contact": {
      "name": "Support API",
      "email": "support@standup-comedy-connect.com"
    }
  },
  "servers": [
    {
      "url": "https://standup-comedy-app.onrender.com/api",
      "description": "Serveur de production"
    },
    {
      "url": "http://localhost:5000/api",
      "description": "Serveur de développement local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/applications": {
      "get": {
        "tags": ["Applications"],
        "summary": "Récupérer toutes les candidatures",
        "description": "Retourne les candidatures selon le rôle de l'utilisateur",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {"type": "string", "enum": ["PENDING", "ACCEPTED", "REJECTED"]}
          },
          {
            "name": "eventId",
            "in": "query",
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des candidatures",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {"type": "string"},
                      "event": {"type": "object"},
                      "comedian": {"type": "object"},
                      "status": {"type": "string", "enum": ["PENDING", "ACCEPTED", "REJECTED"]},
                      "performanceDetails": {
                        "type": "object",
                        "properties": {
                          "duration": {"type": "number"},
                          "description": {"type": "string"},
                          "videoLink": {"type": "string"}
                        }
                      },
                      "message": {"type": "string"},
                      "organizerMessage": {"type": "string"},
                      "createdAt": {"type": "string", "format": "date-time"}
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Applications"],
        "summary": "Créer une nouvelle candidature",
        "description": "Permet à un humoriste de postuler à un événement",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["event", "performanceDetails"],
                "properties": {
                  "event": {"type": "string", "description": "ID de l'événement"},
                  "performanceDetails": {
                    "type": "object",
                    "required": ["duration", "description"],
                    "properties": {
                      "duration": {"type": "number", "minimum": 5, "maximum": 60},
                      "description": {"type": "string", "minLength": 10, "maxLength": 500},
                      "videoLink": {"type": "string", "format": "uri"}
                    }
                  },
                  "message": {"type": "string", "maxLength": 500}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Candidature créée avec succès"
          },
          "400": {
            "description": "Données invalides ou candidature déjà existante"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}